feats by duynd: add voucher service and create multi table-order

+++ src/main/java/com/restarant/backend/model/OrderTotalStatus.java
@@ -4,5 +4,5 @@ public class OrderTotalStatus {
     public static final int ORDERING = 0;
     public static final long WAIT_ACCEPT = 1;
     public static final long ACCEPTED = 2;
-    public static final long PAID = 3;
+    public static final int PAID = 3;
 }
